+++ src/main/java/blackjack/controller/BlackjackGame.java
@@ -13,9 +13,9 @@ import blackjack.model.dealer.Dealer;
 import blackjack.model.player.Player;
 import blackjack.model.player.PlayerName;
 import blackjack.model.player.Players;
-import blackjack.view.Command;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
+import blackjack.view.form.Command;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
