+++ src/main/java/blackjack/view/form/Command.java
@@ -1,4 +1,4 @@
-package blackjack.view;
+package blackjack.view.form;
 
 import java.util.Arrays;
